+++ src/content/injection.ts
@@ -96,13 +96,11 @@ export function bundle(fn: (...args: any[]) => void, deps: Array<(...args: any[]
  * in the page. This also takes care to no create rules with too many selectors
  * for Chrome, see: https://crbug.com/804179
  */
-export function createStylesheet(rules: string[]): string {
+export function createStylesheet(rules: string[], style: string): string {
   const maximumNumberOfSelectors = 1024;
   const parts: string[] = [];
   for (let i = 0; i < rules.length; i += maximumNumberOfSelectors) {
-    parts.push(
-      `${rules.slice(i, i + maximumNumberOfSelectors).join(',')} { display: none!important; }`,
-    );
+    parts.push(`${rules.slice(i, i + maximumNumberOfSelectors).join(',')} { ${style} }`);
   }
   return parts.join('\n');
 }